#29 - removal of finders
