--- server/Routing.Application/Planning/Finders/IRouteFinder.cs
@@ -1,12 +1,0 @@
-﻿using Offroad.Core;
-using Routing.Application.Planning.Intents;
-using Routing.Application.Planning.Profiles;
-using Routing.Domain.Models;
-
-namespace Routing.Application.Planning.Finders
-{
-    public interface IRouteFinder
-    {
-        Task<Result<Trip>> FindRouteAsync(RouteIntent intent, UserRoutingProfile profile, CancellationToken ct);
-    }
-}
