+++ server/Routing.Application/Routes/Commands/RoutesCommands.cs
@@ -4,26 +4,28 @@ using Routing.Application.Contracts.Models;
 using Routing.Application.Contracts.Responses;
 using Routing.Application.Mappings;
 using Routing.Application.Planning.Exceptions;
-using Routing.Application.Planning.Finders;
 using Routing.Domain.Enums;
 using Routing.Domain.Exceptions;
 using Routing.Domain.Models;
 using Routing.Domain.Repositories;
 using Microsoft.Extensions.Logging;
+using Routing.Application.Planning.Candidates.Models;
+using Routing.Application.Planning.Intents;
+using Routing.Application.Planning.Pipelines;
 
 namespace Routing.Application.Routes.Commands
 {
     internal sealed class RoutesCommands : IRoutesCommands
     {
         private readonly ITripRepository _repository;
-        private readonly IRouteFinder _routeFinder;
+        private readonly IPlanningPipeline<RouteIntent, TripCandidate> _planningPipeline;
         private readonly IValidator<PlanRouteRequest> _planValidator;
         private readonly ILogger<RoutesCommands> _logger;
 
-        public RoutesCommands(ITripRepository repository, IRouteFinder routeFinder, IValidator<PlanRouteRequest> planValidator, ILogger<RoutesCommands> logger)
+        public RoutesCommands(ITripRepository repository, IPlanningPipeline<RouteIntent, TripCandidate> planningPipeline, IValidator<PlanRouteRequest> planValidator, ILogger<RoutesCommands> logger)
         {
             _repository = repository;
-            _routeFinder = routeFinder;
+            _planningPipeline = planningPipeline;
             _planValidator = planValidator;
             _logger = logger;
         }
@@ -61,10 +63,18 @@ namespace Routing.Application.Routes.Commands
 
             var intent = request.ToRouteIntent();
             var profile = request.ToUserProfile();
+
             try
             {
-                var routeResult = await _routeFinder.FindRouteAsync(intent, profile, ct);
-                return routeResult.Bind<TripResult>(r => r.ToTripResult(intent));
+                var plans = await _planningPipeline.PlanAsync(intent, profile, ct);
+                var firstPlan = plans.FirstOrDefault(); //for now we only return most suitable route, without any alternatives
+
+                //no plan = no trip :)
+                if (firstPlan is null)
+                    return Error.Validation("Couldn't plan route");
+
+                var trip = Trip.Create("Test route", TripType.Route, firstPlan);
+                return trip.ToTripResult(intent);
             }
             catch (RoutingProviderException ex)
             {
